refactor: rename state helper to storage
